Some updates from trying to install on new macbook
The allowUnfree was because of a complaint about unrar
I suppose it was possible to allow just that one, but I don't really
care so I'd rather set it globally. Kind of confused as to why setting
these properties on nixpkgs instead of pkgs or unstable works, but oh
well.